minor change... Charles, ne t'inquiètes pas... Pourquoi ai-je transformé les long en Integer ? - type "objet" plus facile à manipuler pour ses méthodes et les conversions - largement suffisant pour nos besoins donc mieux pour l'espace mémoire occupé et je sais que ça t'importe. Non? Bon weekend!!

+++ src/main/java/fr/eni/filmotheque/ihm/FilmsController.java
@@ -6,10 +6,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import ch.qos.logback.core.recovery.ResilientSyslogOutputStream;
 import fr.eni.filmotheque.bll.FilmsService;
 import fr.eni.filmotheque.bo.Film;
 
@@ -45,7 +43,7 @@ public class FilmsController
 	}
 	
 	@GetMapping({"/movies"})
-	public String movies(@RequestParam long id, Model model)
+	public String movies(@RequestParam Integer id, Model model)
 	{	
 		model.addAttribute("movie",filmsService.getFilmById(id));
